requirements.txt
h11~=0.14.0
numpy~=2.2.4
typing_extensions~=4.13.1
pillow~=11.1.0
pip~=24.2
filelock~=3.18.0
pytz~=2025.2
rich~=14.0.0
Pygments~=2.19.1
requests~=2.32.3
colorama~=0.4.6
PyYAML~=6.0.2
sniffio~=1.3.1
ffmpy~=0.5.0
click~=8.1.8
httpcore~=1.0.7
idna~=3.10
httpx~=0.28.1
mdurl~=0.1.2
typer~=0.15.2
shellingham~=1.5.4
fsspec~=2025.3.2
gradio_client~=1.8.0
gradio~=5.24.0
tomlkit~=0.13.2
fastapi~=0.115.12
starlette~=0.46.1
anyio~=4.9.0
orjson~=3.10.16
groovy~=0.1.2
Jinja2~=3.1.6
MarkupSafe~=3.0.2
pydub~=0.25.1
pandas~=2.2.3
pydantic~=2.11.3
packaging~=24.2
uvicorn~=0.34.0
pydantic_core~=2.33.1
aiofiles~=24.1.0
safehttpx~=0.1.6
urllib3~=2.3.0
python-dateutil~=2.9.0.post0
pyserial~=3.5
certifi~=2025.1.31
websockets~=15.0.1
six~=1.17.0
tqdm~=4.67.1
python-multipart~=0.0.20